moving transfer_form into a groups module
