+++ test/components/groups/transfer_in_activity_component_test.rb
@@ -13,7 +13,7 @@ module Groups
 
     test 'transfer group into group activity' do
       group2 = groups(:group_two)
-      transfer_form = ::TransferForm.new({ new_namespace_id: @group.id }
+      transfer_form = ::Groups::TransferForm.new({ new_namespace_id: @group.id }
       .merge(group_id: group2.id, group_name: group2.name, group_path: group2.path))
 
       ::Groups::TransferService.new(group2, @user, transfer_form).execute(@group)
@@ -43,7 +43,7 @@ module Groups
     test 'transfer group from an existing group namespace into group activity' do
       group2 = groups(:group_two)
       subgroup = groups(:subgroup1)
-      transfer_form = ::TransferForm.new({ new_namespace_id: group2.id }
+      transfer_form = ::Groups::TransferForm.new({ new_namespace_id: group2.id }
       .merge(group_id: subgroup.id, group_name: subgroup.name, group_path: subgroup.path))
 
       ::Groups::TransferService.new(subgroup, @user, transfer_form).execute(group2)
@@ -74,7 +74,7 @@ module Groups
     test 'transfer group from an existing group namespace into group then delete transferred group activity' do
       group2 = groups(:group_two)
       subgroup = groups(:subgroup1)
-      transfer_form = ::TransferForm.new({ new_namespace_id: group2.id }
+      transfer_form = ::Groups::TransferForm.new({ new_namespace_id: group2.id }
       .merge(group_id: subgroup.id, group_name: subgroup.name, group_path: subgroup.path))
 
       ::Groups::TransferService.new(subgroup, @user, transfer_form).execute(group2)
